refactor(editor): slim the shell — EditorHeader/ToolRail/PanelChrome/CropPanel (Phase 4)
Pull the presentational shell out of EditorOverlay into four components:

- EditorHeader  — top bar (source label + edited/dims readout + Export/Close)
- ToolRail      — the vertical tool switcher (absorbs the ToolTab helper)
- PanelChrome   — the bordered right-panel container + title/badge header
- CropPanel     — the crop/transform panel body (absorbs AspectButton +
                  FooterButton + getAspectPreviewBox), so the crop tool now
                  matches the <TextPanel>/<FramePanel>/... shape

All state stays in EditorOverlay; the new components are pure/presentational
and receive values + handlers as props. The dims-readout IIFE and the
tool-switch depth-reset are hoisted to plain locals (dimsLabel, selectTool)
with identical logic. Dropped the now-unused lucide-react icon imports plus
ASPECT_PRESETS / getAspectRatio / MAX_IMAGE_ZOOM.

Behavior-preserving: golden e2e 03/04/06/18/19/20 green (23 passed).

EditorOverlay 1195 → 986 lines.
